Move createCharacterForm
