+++ the-wanderers/scripts/helpers.js
@@ -1,5 +1,6 @@
 import { Character, ageArray, hungerArray, moraleArray, healthArray, weaponArray } from './character.js';
 import { playTurn } from './game.js';
+import Party from './party.js';
 
 const context = {
     gameParty: null
@@ -893,4 +894,333 @@ styleSelector.addEventListener("change", () => {
     document.querySelector("link[rel='stylesheet']").href = styleSelector.value;
 });
 
-export { context, setGameParty, addItemToInventory, getEvent, updateStatBars, food, medical, addEvent, getName, posTraits, negTraits, updateRelationships, updateFoodButtons, updateMedicalButtons, foundFriend, foundEnemy, foundFood, foundMedical, foundWeapon, checkDeathEffects, updateInteractionButtons };
+async function createCharacterForm() {
+    const formDiv = document.getElementById('characters');
+    const precis = document.createElement('p');
+
+    precis.textContent = 'Create a character to start the game.';
+    formDiv.appendChild(precis);
+
+    const form = document.createElement('form');
+    const response = await fetch('https://randomuser.me/api/?nat=au,br,ca,ch,de,dk,es,fi,fr,gb,ie,in,mx,nl,no,nz,rs,tr,ua,us');
+    const data = await response.json();
+    const firstName = getName(data);
+
+    const nameLabel = document.createElement('label');
+    nameLabel.textContent = 'Name: ';
+    const nameInput = document.createElement('input');
+    nameInput.type = 'text';
+    nameInput.value = firstName;
+    nameLabel.appendChild(nameInput);
+    form.appendChild(nameLabel);
+
+    const ageLabel = document.createElement('label');
+    ageLabel.textContent = 'Age: ';
+    const ageInput = document.createElement('select');
+    for (let i = 0; i < ageArray.length; i++) {
+        const option = document.createElement('option');
+        option.value = i;
+        option.textContent = ageArray[i];
+        ageInput.appendChild(option);
+      }
+    ageInput.selectedIndex = Math.floor(Math.random() * ageArray.length);
+    ageLabel.appendChild(ageInput);
+    form.appendChild(ageLabel);
+
+    const posTraitsLabel = document.createElement('label');
+    posTraitsLabel.textContent = 'Positive Trait: ';
+    const posTraitsSelect = document.createElement('select');
+    for (const trait of posTraits) {
+      const option = document.createElement('option');
+      option.value = trait[0];
+      option.textContent = trait[0];
+      posTraitsSelect.appendChild(option);
+    }
+    posTraitsSelect.selectedIndex = Math.floor(Math.random() * posTraits.length);
+    posTraitsLabel.appendChild(posTraitsSelect);
+    form.appendChild(posTraitsLabel);
+
+    const posTraitsDescription = document.createElement('p');
+    const posEffectsList = document.createElement('ul');
+    let posEffect1 = document.createElement('li');
+    posEffect1.textContent = posTraits[posTraitsSelect.selectedIndex][1];
+    posEffectsList.appendChild(posEffect1);
+    let posEffect2 = document.createElement('li');
+    posEffect2.textContent = posTraits[posTraitsSelect.selectedIndex][2];
+    posEffectsList.appendChild(posEffect2);
+    posTraitsDescription.appendChild(posEffectsList);
+    form.appendChild(posTraitsDescription);
+
+    const negTraitsLabel = document.createElement('label');
+    negTraitsLabel.textContent = 'Negative Trait: ';
+    const negTraitsSelect = document.createElement('select');
+    for (const trait of negTraits) {
+      const option = document.createElement('option');
+      option.value = trait[0];
+      option.textContent = trait[0];
+      negTraitsSelect.appendChild(option);
+    }
+    negTraitsSelect.selectedIndex = Math.floor(Math.random() * negTraits.length);
+    negTraitsLabel.appendChild(negTraitsSelect);
+    form.appendChild(negTraitsLabel);
+
+    const negTraitsDescription = document.createElement('p');
+    const negEffectsList = document.createElement('ul');
+    let negEffect1 = document.createElement('li');
+    negEffect1.textContent = negTraits[negTraitsSelect.selectedIndex][1];
+    negEffectsList.appendChild(negEffect1);
+    let negEffect2 = document.createElement('li');
+    negEffect2.textContent = negTraits[negTraitsSelect.selectedIndex][2];
+    negEffectsList.appendChild(negEffect2);
+    negTraitsDescription.appendChild(negEffectsList);
+    form.appendChild(negTraitsDescription);
+
+    // Avatar creation section
+    const avatarSection = document.createElement('div');
+    avatarSection.innerHTML = '<p>Create your avatar:</p>';
+
+    // Skin selection
+    const skinLabel = document.createElement('label');
+    skinLabel.textContent = 'Skin: ';
+    const skinSelect = document.createElement('select');
+    const skinImages = ['skin_dark.png', 'skin_dark-mid.png', 'skin_mid.png', 'skin_light-mid.png', 'skin_light.png'];
+    skinImages.forEach(skin => {
+        const option = document.createElement('option');
+        option.value = skin;
+        const skinText = skin.split('.')[0];
+        option.textContent = skinText.split('_')[1];
+        skinSelect.appendChild(option);
+    });
+    skinSelect.selectedIndex = Math.floor(Math.random() * skinImages.length);
+    skinLabel.appendChild(skinSelect);
+    avatarSection.appendChild(skinLabel);
+
+    // Hair style selection
+    const hairStyleLabel = document.createElement('label');
+    hairStyleLabel.textContent = 'Hair Style: ';
+    const hairStyleSelect = document.createElement('select');
+    const hairStyleImages = ['hair_long-curly', 'hair_long-straight', 'hair_short-fluffy', 'hair_short-straight'];
+    hairStyleImages.forEach(hairStyle => {
+        const option = document.createElement('option');
+        option.value = hairStyle;
+        option.textContent = hairStyle.split('_')[1].replace('-',' and ');
+        hairStyleSelect.appendChild(option);
+    });
+    hairStyleSelect.selectedIndex = Math.floor(Math.random() * hairStyleImages.length);
+    hairStyleLabel.appendChild(hairStyleSelect);
+    avatarSection.appendChild(hairStyleLabel);
+
+    // Hair colour selection
+    const hairColourLabel = document.createElement('label');
+    hairColourLabel.textContent = 'Hair Colour: ';
+    const hairColourSelect = document.createElement('select');
+    const hairColourImages = ['blonde.png', 'ginger.png', 'brown.png', 'red.png', 'black.png'];
+    hairColourImages.forEach((hairColour, index) => {
+        const option = document.createElement('option');
+        option.value = hairColour;
+        // grab the name of the file before the .png
+        option.textContent = hairColourImages[index].split('.')[0];
+        hairColourSelect.appendChild(option);
+    });
+    hairColourSelect.selectedIndex = Math.floor(Math.random() * hairColourImages.length);
+    hairColourLabel.appendChild(hairColourSelect);
+    avatarSection.appendChild(hairColourLabel);
+
+    // shirt style selection
+    const shirtStyleLabel = document.createElement('label');
+    shirtStyleLabel.textContent = 'Shirt Style: ';
+    const shirtStyleSelect = document.createElement('select');
+    const shirtStyleImages = ['shirt_hoodie', 'shirt_jacket', 'shirt_scarf', 'shirt_vest'];
+    shirtStyleImages.forEach(shirtStyle => {
+        const option = document.createElement('option');
+        option.value = shirtStyle;
+        option.textContent = shirtStyle.split('_')[1];
+        shirtStyleSelect.appendChild(option);
+    });
+    shirtStyleSelect.selectedIndex = Math.floor(Math.random() * shirtStyleImages.length);
+    shirtStyleLabel.appendChild(shirtStyleSelect);
+    avatarSection.appendChild(shirtStyleLabel);
+
+    // shirt colour selection
+    const shirtColourLabel = document.createElement('label');
+    shirtColourLabel.textContent = 'Shirt Colour: ';
+    const shirtColourSelect = document.createElement('select');
+    const shirtColourImages = ['red.png', 'yellow.png', 'green.png', 'blue.png'];
+    shirtColourImages.forEach((shirtColour, index) => {
+        const option = document.createElement('option');
+        option.value = shirtColour;
+        option.textContent = shirtColourImages[index].split('.')[0];
+        shirtColourSelect.appendChild(option);
+    });
+    shirtColourSelect.selectedIndex = Math.floor(Math.random() * shirtColourImages.length);
+    shirtColourLabel.appendChild(shirtColourSelect);
+    avatarSection.appendChild(shirtColourLabel);
+
+    // Avatar preview container
+    const avatarPreviewContainer = document.createElement('div');
+    avatarPreviewContainer.className = 'avatar';
+    avatarPreviewContainer.style.position = 'relative';
+
+    const avatarPreview = document.createElement('div');
+    avatarPreview.className = 'avatarSprite';
+
+    // Skin preview
+    const skinPreview = avatarPreview.appendChild(document.createElement('img'));
+    skinPreview.src = "images/skin/" + skinSelect.value;
+    skinPreview.alt = nameInput.value + '\'s skin sprite.';
+    avatarPreview.appendChild(skinPreview);
+
+    // Hair preview
+    const hairPreview = avatarPreview.appendChild(document.createElement('img'));
+    hairPreview.src = "images/hair/" + hairStyleSelect.value + '_' + hairColourSelect.value;
+    hairPreview.alt = nameInput.value + '\'s hair sprite. . Their hair is ';
+    switch (hairStyleSelect.value) {
+        case 'short1':
+            hairPreview.alt += 'short and straight.';
+            break;
+        case 'short2':
+            hairPreview.alt += 'short and fluffy.';
+            break;
+        case 'long1':
+            hairPreview.alt += 'long and straight.';
+            break;
+        case 'long2':
+            hairPreview.alt += 'long and curly.';
+            break;
+    }
+    avatarPreview.appendChild(hairPreview);
+
+    // shirt preview
+    const shirtPreview = avatarPreview.appendChild(document.createElement('img'));
+    shirtPreview.src = "images/shirts/" + shirtStyleSelect.value + '_' + shirtColourSelect.value;
+    shirtPreview.alt = nameInput.value + '\'s shirt sprite';
+    switch (shirtStyleSelect.value) {
+      case "images/shirts/shirt1.png":
+        shirtPreview.alt += "hoodie.";
+        break;
+      case "images/shirts/shirt2.png":
+        shirtPreview.alt += "vest.";
+        break;
+      case "images/shirts/shirt3.png":
+        shirtPreview.alt += "jacket.";
+        break;
+      case "images/shirts/shirt4.png":
+        shirtPreview.alt += "scarf.";
+        break;
+    }avatarPreview.appendChild(shirtPreview);
+
+    avatarPreviewContainer.appendChild(avatarPreview);
+    avatarSection.appendChild(avatarPreviewContainer);
+
+
+    // Event listeners to update previews
+    posTraitsSelect.addEventListener('change', () => {
+        const posTraitIndex = posTraits.findIndex(trait => trait[0] === posTraitsSelect.value);
+        const effectsList = document.createElement('ul');
+        let effect1 = document.createElement('li');
+        effect1.textContent = posTraits[posTraitIndex][1];
+        effectsList.appendChild(effect1);
+        let effect2 = document.createElement('li');
+        effect2.textContent = posTraits[posTraitIndex][2];
+        effectsList.appendChild(effect2);
+        posTraitsDescription.innerHTML = '';
+        posTraitsDescription.appendChild(effectsList);
+    });
+
+    negTraitsSelect.addEventListener('change', () => {
+        const negTraitIndex = negTraits.findIndex(trait => trait[0] === negTraitsSelect.value);
+        const effectsList = document.createElement('ul');
+        let effect1 = document.createElement('li');
+        effect1.textContent = negTraits[negTraitIndex][1];
+        effectsList.appendChild(effect1);
+        let effect2 = document.createElement('li');
+        effect2.textContent = negTraits[negTraitIndex][2];
+        effectsList.appendChild(effect2);
+        negTraitsDescription.innerHTML = '';
+        negTraitsDescription.appendChild(effectsList);
+    });
+
+    skinSelect.addEventListener('change', () => {
+        skinPreview.src = "images/skin/" + skinSelect.value;
+    });
+
+    hairStyleSelect.addEventListener('change', () => {
+        hairPreview.src = "images/hair/" + hairStyleSelect.value + '_' + hairColourSelect.value;
+    });
+
+    hairColourSelect.addEventListener('change', () => {
+        hairPreview.src = "images/hair/" + hairStyleSelect.value + '_' + hairColourSelect.value;
+    });
+
+    shirtStyleSelect.addEventListener('change', () => {
+        shirtPreview.src = "images/shirts/" + shirtStyleSelect.value + '_' + shirtColourSelect.value;
+    });
+
+    shirtColourSelect.addEventListener('change', () => {
+        shirtPreview.src = "images/shirts/" + shirtStyleSelect.value + '_' + shirtColourSelect.value;
+    });
+
+    form.appendChild(avatarSection);
+
+    const submitButton = document.createElement('button');
+    submitButton.textContent = 'Start Game';
+    submitButton.addEventListener('click', () => {
+        const name = nameInput.value;
+        const age = ageInput.value;
+        const posTrait = posTraitsSelect.value;
+        const negTrait = negTraitsSelect.value;
+        const skin = "images/skin/" + skinSelect.value;
+        const hair = "images/hair/" + hairStyleSelect.value + '_' + hairColourSelect.value;
+        const shirt = "images/shirts/" + shirtStyleSelect.value + '_' + shirtColourSelect.value;
+        const character = new Character(name, age, posTrait, negTrait, skin, hair, shirt);
+        formDiv.innerHTML = "";
+        startGame();
+        const playTurnButton = document.createElement('button');
+        playTurnButton.id = 'playTurnButton';
+        playTurnButton.textContent = 'Play Turn 1';
+        playTurnButton.addEventListener('click', () => {
+            playTurn();
+        });
+        document.getElementById('gameButtons').appendChild(playTurnButton);
+        context.gameParty.addCharacter(character);
+        character.createCharacter();
+        character.updateCharacter();
+        updateStatBars(character);
+
+        //unhide the buttons div
+        const buttonsDiv = document.getElementById('buttons');
+        buttonsDiv.style.display = 'block';
+
+        //unhide the events div
+        const eventsDiv = document.getElementById('content');
+        eventsDiv.style.display = 'flex';
+
+        addEvent(`A new illness has swept the world and the infected have begun to rise from the dead. The world is ending, but ${character.name}'s life doesn't have to just yet.`)
+        document.documentElement.scrollTop = 0;
+        document.body.scrollTop = 0;
+    });
+    form.appendChild(submitButton);
+
+    // Set focus on the name input field
+    nameInput.focus();
+
+    // Submit the form when Enter key is pressed in the name input field
+    nameInput.addEventListener('keydown', function(event) {
+    if (event.key === 'Enter') {
+      event.preventDefault();
+      form.submit();
+    }
+    });
+
+    formDiv.appendChild(form);
+}
+
+async function startGame() {
+    let gameParty;
+    if (!context.gameParty) {
+        gameParty = new Party();
+    }
+    setGameParty(gameParty);
+}
+
+export { context, setGameParty, addItemToInventory, getEvent, updateStatBars, food, medical, addEvent, getName, posTraits, negTraits, updateRelationships, updateFoodButtons, updateMedicalButtons, foundFriend, foundEnemy, foundFood, foundMedical, foundWeapon, checkDeathEffects, updateInteractionButtons, createCharacterForm };
